Parameterized the xml file location
In the before starting file it suggested that we should update the
scripts to specify where the xml file lives. I have done that. Recommend
copying the xml file to the root of a local or mapped drive to expedite
running the scripts. Input now looks like: 01-Set-SQLAlias.ps1 2016 R: